Replace use of redux toolkit createSlice with createReducer and a centralized action creation in actiontypes.js
This is because the actions from createSlice resulted in confusing
imports that look messy and are hard to use.

Using createSlice was particularily messy for those reducers handling
responses from REST calls, because the initial action that triggered
the REST call needed to be created outside of the createSlice() using
createAction().

So I decided it was better to create all actions in the same matter,
in the same place, using createAction()