Change bootstrap breakpoint from "md" to "lg" to get "image expand on landscape" on modern high-res phones
The image expanded to fill available space in chrome mobile emulator
when I "tilted" it.

But the image didn't expand on an Xperia 10 III or a recent mobile phone.

I eventually figured out that this was caused by the phones having
hight resolution displays (the Xperia has a 1080x2520 display, which
is an "lg" display in bootstrap terms).