Disable sonarqube/sonarcloud rule java:S3626 since it encourages bad practice
Martin Fowler in "Recfactoring" says, and I agree, that in a function
you should check if you have anything to do in that function as early
as possible and then get out of that function.

Rule java:S3626 says that you should return from the end of the
function after having matched none of the rules, in the case of a
no-op.

And that's more error prone and harder to read.

So begone with java:S3626!