Remove web context from saved original request URL when logging in
Also replace some check-for-null logic in JAX-RS resource with
throwing exception.

Note that the reason for logging on the Login resource, instead of
putting the information in the thrown exceptions, is as to not
unintentionally reveal login information to the caller.

The exception causing the problem is logged but not made part of the
thrown exception.